Min length size_t

+++ include/stellar/options/eps_match_options.hpp
@@ -10,7 +10,7 @@ struct EPSMatchOptions
 {
     stellar::utils::fraction epsilon{5, 100}; // maximal error rate
     double numEpsilon{0.05};
-    unsigned minLength{100}; // minimal length of an epsilon-match
+    size_t minLength{100}; // minimal length of an epsilon-match
 };
 
 } // namespace stellar
